check for returned runs

+++ master/internal/experiment/bulk_action.go
@@ -800,27 +800,29 @@ func MoveExperiments(
 			return nil, err
 		}
 
-		var sourceProjectID int
-		err = tx.NewSelect().Table("runs").Column("project_id").
-			Where("id IN (?)", bun.In(runIDs)).
-			Limit(1).
-			Scan(ctx, &sourceProjectID)
-		if err != nil {
-			return nil, err
-		}
+		if len(runIDs) > 0 {
+			var sourceProjectID int
+			err = tx.NewSelect().Table("runs").Column("project_id").
+				Where("id IN (?)", bun.In(runIDs)).
+				Limit(1).
+				Scan(ctx, &sourceProjectID)
+			if err != nil {
+				return nil, err
+			}
 
-		if _, err = tx.NewUpdate().Table("runs").
-			Set("project_id = ?", destinationProjectID).
-			Where("runs.experiment_id IN (?)", bun.In(validIDs)).
-			Exec(ctx); err != nil {
-			return nil, fmt.Errorf("updating run's project IDs: %w", err)
-		}
+			if _, err = tx.NewUpdate().Table("runs").
+				Set("project_id = ?", destinationProjectID).
+				Where("runs.experiment_id IN (?)", bun.In(validIDs)).
+				Exec(ctx); err != nil {
+				return nil, fmt.Errorf("updating run's project IDs: %w", err)
+			}
 
-		if err = db.AddProjectHparams(ctx, tx, int(destinationProjectID), runIDs); err != nil {
-			return nil, err
-		}
-		if err = db.RemoveOutdatedProjectHparams(ctx, tx, sourceProjectID); err != nil {
-			return nil, err
+			if err = db.AddProjectHparams(ctx, tx, int(destinationProjectID), runIDs); err != nil {
+				return nil, err
+			}
+			if err = db.RemoveOutdatedProjectHparams(ctx, tx, sourceProjectID); err != nil {
+				return nil, err
+			}
 		}
 
 		for _, acceptID := range acceptedIDs {
